// src/antrian-pasien/dtos/recap-antrian.dto.ts
export class RecapAntrianPasiensPerdayRawDto {
  user_id: number;
  total: number;
  perday: Date;
}
export class RecapAntrianPasiensPermonthRawDto {
  user_id: number;
  total: number;
  permonth: Date;
}
export class RecapAntrianPasiensPeryearRawDto {
  user_id: number;
  total: number;
  peryear: Date;
}